192/396:

+++ ft_is2.c
@@ -10,10 +10,10 @@ void	ft_is_o(int *i, t_en *e)
 	e->c == 0 ? nb = va_arg(e->a, int) : 0;
 	e->c == 'H' ? nb = (unsigned char)va_arg(e->a, int) : 0;
 	e->c == 'h' ? nb = (unsigned short int)va_arg(e->a, int) : 0;
-	e->c == 'l' ? nb = (unsigned long int)va_arg(e->a, int) : 0;
-	e->c == 'L' ? nb = (unsigned long long int)va_arg(e->a, int) : 0;
+	e->c == 'l' ? nb = (unsigned long int)va_arg(e->a, unsigned long long int) : 0;
+	e->c == 'L' ? nb = (unsigned long long int)va_arg(e->a, unsigned long long int) : 0;
 	e->c == 'j' ? nb = (int)va_arg(e->a, int) : 0;
-	e->c == 'z' ? nb = (size_t)va_arg(e->a, int) : 0;
+	e->c == 'z' ? nb = (size_t)va_arg(e->a, size_t) : 0;
 	e->print = ft_strlen(ft_itoa_base(nb, 8));
 	if (e->m == 0)
 	{
